+++ src/components/modules/Reviews/defaultReviews.json
@@ -1,0 +1,30 @@
+[
+  {
+    "_id": 1,
+    "avatar": "../../../../img/reviews/review-1.png",
+    "name": "Оксана К.",
+    "date": "2022-12-10T12:25:44.131Z",
+    "comments": "Дівчата, ці тістечка посто бомба-ракєта!!! Нарешті дійшов мій марципан, і я одразу приготувала, тому тепер я в графіку і в ногу з іншими учасниками курсу."
+  },
+  {
+    "_id": 2,
+    "avatar": "../../../../img/reviews/review-2.png",
+    "name": "Ніка",
+    "date": "2022-09-15T12:25:44.131Z",
+    "comments": "Дівчата, ці тістечка посто бомба-ракєта!!! Нарешті дійшов мій марципан, і я одразу приготувала, тому тепер я в графіку і в ногу з іншими учасниками курсу."
+  },
+  {
+    "_id": 3,
+    "avatar": "../../../../img/reviews/review-3.png",
+    "name": "Олександра",
+    "date": "2022-08-23T12:25:44.131Z",
+    "comments": "Дівчата, ці тістечка посто бомба-ракєта!!! Нарешті дійшов мій марципан, і я одразу приготувала, тому тепер я в графіку і  в ногу з іншими учасниками курсу."
+  },
+  {
+    "_id": 4,
+    "avatar": "../../../../img/reviews/review-4.png",
+    "name": "Олександра",
+    "date": "2022-05-06T12:25:44.131Z",
+    "comments": "Дівчата, ці тістечка посто бомба-ракєта!!! Нарешті дійшов мій марципан, і я одразу приготувала, тому тепер я в графіку і в ногу з іншими учасниками курсу."
+  }
+]
