+++ src/components/modules/Reviews/index.js
@@ -1,1 +1,1 @@
-export { Reviews } from './Reviews';
+export { Reviews } from './Review';
